+++ doc/version_history.rst
@@ -6,6 +6,22 @@
 Version History
 ###############
 
+v0.3.0
+======
+
+* Moved all device reply validating code from ts.ess.controller to ts.ess.common.
+* Moved all sensors code from ts.ess.controller to ts.ess.common.
+* Moved code to determine what sensor is connected from ts.ess.controller to ts.ess.common.
+* Moved BaseDevice and MockDevice from ts.ess.controller to ts.ess.common.
+* Added a unit test for the config schema.
+* Moved most of the command handler code and the socket server unit test from ts.ess.controller to ts.ess.common.
+* Added tests for all supported devices in the test class for the mock control handler.
+
+Requires:
+
+* ts_tcpip 0.3
+* ts_utils 1.0
+
 v0.2.0
 ======